Add MIT License
This pull request adds the MIT License to the project. The purpose of this change is to clearly define the terms of use for the project and provide clarity for potential contributors and users.

The scope of the changes includes the addition of a LICENSE file to the root directory of the project. The contents of the LICENSE file contain the text of the MIT License, which grants users the right to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the software.

By adding the MIT License, we hope to encourage wider adoption of the project and make it easier for developers to use and contribute to the codebase. This change does not affect the functionality of the project, and all existing code remains unchanged.